requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
pydantic==2.8.2
pyyaml==6.0.2
cryptography==43.0.3
jinja2==3.1.4
python-multipart==0.0.12
